refactor docker image spec generation
