+++ pkg/plugins/autodiscovery/dockercompose/utils_test.go
@@ -31,17 +31,20 @@ func TestGetDockerComposeSpec(t *testing.T) {
 
 	expectedDockerComposeSpec := DockerComposeSpec{
 		Services: map[string]Service{
-			"agent": {
-				Image: "ghcr.io/updatecli/updatemonitor:v0.1.0",
-			},
-			"front": {
-				Image: "ghcr.io/updatecli/updatemonitor-ui:v0.1.1",
+			"agent-2": {
+				Image: "ghcr.io/updatecli/updatemonitor@sha256:2b3436753ba2c671c382777d8980ed6ea7350851e71bff0326dc776c0c9e2334",
 			},
 			"mongodb": {
 				Image: "mongo:6.0.2",
 			},
+			"agent": {
+				Image: "ghcr.io/updatecli/updatemonitor:v0.1.0",
+			},
 			"server": {
 				Image: "ghcr.io/updatecli/updatemonitor:v0.1.0",
+			},
+			"front": {
+				Image: "ghcr.io/updatecli/updatemonitor-ui:v0.1.1",
 			},
 			"traefik": {
 				Image: "traefik:v2.9",
